+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased]
+### Added
+- Added time tagging to channels
+
+
 ## [0.3.7]
 ### Added
 - Added `WaveformPulse` to allow for pre-defined waveforms.